+++ rootbox/cli/cmd_run.py
@@ -8,11 +8,10 @@ import typer
 
 from ..download import download_image
 from ..images import parse_image_url
+from ..mount_checker import MountChecker
 from ..rootfs import prepare_rootfs
 from ..shell.execute import execute
 from ..unshare import CLONE_NEWNET, unshare
-
-MOUNT_DIR: Path = None
 
 
 def run(
@@ -21,27 +20,31 @@ def run(
     no_net: bool = typer.Option(
         False, "--no-network", "-N", help="Disable network in the container"
     ),
+    ram_disk_size: Optional[int] = typer.Option(
+        1, "--ram-disk", "-r", help="Size of the ram disk (GBs)"
+    ),
     command: Optional[str] = typer.Argument(None, help="Command to be run"),
     only_from_cache: bool = typer.Option(False, "--only-from-cache"),
     tar_file: Optional[Path] = typer.Option(None, "--tar-file", "-t"),
 ):
-    global MOUNT_DIR
     if no_shell and command == "/bin/sh":
         raise typer.BadParameter("--no-shell was provided but no command was given")
 
     image = parse_image_url(image_name)
     if isinstance(image, Path):
         image_fname = image
+        image_prompt_name = f"file:{image.name}"
     else:
         image_fname = download_image(image, only_from_cache=only_from_cache)
-    mount_dir = prepare_rootfs(image_fname, in_memory=True, perform_chroot=True)
+        image_prompt_name = image_name
+    mount_dir = prepare_rootfs(image_fname, ram_disk_size, perform_chroot=True)
+
     if no_net:
         unshare(CLONE_NEWNET)
-    execute(image_name, mount_dir, command, use_shell=not no_shell)
-    # tar_file = Path("/hostroot") / tar_file.relative_to("/")
+    execute(image_prompt_name, mount_dir, command, use_shell=not no_shell)
+    MountChecker.read_mounts()
     os.chroot("/host_root")
     if tar_file:
-        MOUNT_DIR = mount_dir
         with tarfile.open(tar_file, "w:gz") as tar:
             tar.add(
                 mount_dir, arcname="./", filter=filter_out_other_mounts, recursive=True
@@ -49,22 +52,5 @@ def run(
 
 
 def filter_out_other_mounts(tarinfo: tarfile.TarInfo):
-    global MOUNT_DIR
-    filename = MOUNT_DIR.joinpath(tarinfo.name)
-    if is_on_another_filesystem(filename):
-        return None
-    return tarinfo
-
-
-def is_on_another_filesystem(path: Path):
-    global MOUNT_DIR
-    if path.is_symlink():
-        return False
-    if path == MOUNT_DIR:
-        return False
-    st = os.stat(path)
-    current_dev = st.st_dev
-    st = os.stat(path.parent)
-    if st.st_dev != current_dev:
-        return True
-    return False
+    if MountChecker.is_on_mount_dir(tarinfo.name):
+        return tarinfo