+++ rootbox/shell/execute.py
@@ -13,11 +13,13 @@ def execute(
     """Execute the command in the container as a shell command
     or an interactive shell if no command is provided"""
     if command is None:
-        command = "/bin/sh"
-    is_interactive = command == "/bin/sh"
-    if is_interactive:
+        use_shell = False
+        if Path("/bin/bash").exists():
+            command = "/bin/bash"
+        elif Path("/bin/sh").exists():
+            command = "/bin/sh"
+        else:
+            raise Exception("No shell found in the container")
         os.environ["PS1"] = f"\e[0;94m(rbox {image_name} \w)$ \e[m"
         print_system_info()
-    if command == "/bin/sh":
-        use_shell = False
     subprocess.call(command, shell=use_shell)