+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to the currrent release will be documented in this file.
 
 ---
 
-## 0.0.8
+## 0.0.9
 
 ### Added
-- Option `--save-to-tar` to save image file after run
+- Option `--save-to-tar` to save image file after `run`
+- Support images stored in  local .tar.gz files
+- Automatic shell detection from image contents (bash, sh)
 
 ### Changed
 - TODO: Add --save-to-tar option on `rootbox run`